Change feasibility tolerance in LPSolver

+++ src/main/kotlin/core/environment/EnvironmentParameters.kt
@@ -2,5 +2,5 @@ package core.environment
 
 data class EnvironmentParameters(
     val nCloud: Int,
-    val tRx: Int,
+    val tRx: Double,
 )
